requirements.txt
langchain~=0.3.27
langchain-community~=0.3.29
langchain-core~=0.3.75
langchain-huggingface~=0.3.1
langchain-milvus~=0.2.1
langchain-openai~=0.3.32
langchain-text-splitters~=0.3.11
pydantic~=2.11.7
gradio~=5.44.1
langgraph~=0.4.0
pymilvus~=2.6.0
numpy~=1.26.4
pillow~=10.4.0
opencv-python-headless~=4.10.0.82
torch~=2.8.0
torchvision~=0.23.0
scikit-learn~=1.5.0
scipy~=1.15.3
sentence-transformers
typeguard